requirements.txt
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
crypto==1.4.1
cryptography==36.0.1
Django==2.2.5
django-filter==21.1
django-inline-media==1.4.2
django-summernote==0.8.20.0
djangorestframework==3.13.1
djangorestframework-api-key==2.1.0
idna==3.3
importlib-metadata==4.10.1
intasend-python==1.0.1
Markdown==3.3.6
Naked==0.1.31
packaging==21.3
pycparser==2.21
pycrypto==2.6.1
pyOpenSSL==21.0.0
pyparsing==3.0.7
pytz==2021.3
PyYAML==6.0
requests==2.27.1
shellescape==3.8.1
six==1.16.0
sorl-thumbnail==12.7.0
tagging==0.2.1
typing_extensions==4.0.1
urllib3==1.26.8
webencodings==0.5.1
zipp==3.7.0
